Moved processing of sets into a helper function. Also backup handles files better now

+++ data_helpers.py
@@ -297,6 +297,9 @@ def write_bad_comparisons(all_data: pd.DataFrame,
 
 def backup_file(file_name: str,
                 file_directory: str = "./"):
+    if not os.path.exists(f"{file_directory}{file_name}"):
+        print(f"Failed to back up {file_name} becuase the file does not exist yet.")
+        return
     try:
         print(f"Backing up {file_name}...", end="", flush=True)
         shutil.copyfile(f"{file_directory}{file_name}", f"./backups/[{datetime.datetime.today().strftime('%Y%m%d')}] {file_name}")
@@ -305,4 +308,35 @@ def backup_file(file_name: str,
         print("backup directory does not exist. Making /backups/")
         os.mkdir("./backups")
         backup_file(file_name)
+
+
+def process_sets(sets: pd.DataFrame,
+                 new_data: pd.DataFrame):
+    good_rows = []
+    good_comparisons = []
+    bad_comparisons = []
+    
+    for _, row in sets.iterrows():
+        date, nuance, hair_type = row
+        filtered_data = filter_for_group(new_data,
+                                         date,
+                                         nuance,
+                                         hair_type).reset_index(drop=True)
         
+        if filtered_data.shape[0] <= 1:  # No comparisons can be made if there is only 1 data point in the set.
+            bad_comparisons.append(filtered_data)
+        elif sum(filtered_data["STD"]) < 1:  # No standards in the set
+            bad_comparisons.append(filtered_data)
+        elif sum(filtered_data["STD"]) >= 2:  # Too many standards for comparisons.
+            #TODO Narrow filter band
+            pass
+        else:
+            good_rows.append(filtered_data)
+            standard = filtered_data.loc[filtered_data["STD"] == True]
+            for index in filtered_data.index:
+                if index == standard.index:  # Skips testing standards against themselves
+                    continue
+                comparison = filtered_data.loc[index:index+1]
+                good_comparisons.append(report_comparison(standard, comparison))
+    
+    return good_rows, good_comparisons, bad_comparisons
